Fix for waves
There was an issue with the waves using an interval to spawn the waves which could be off by a couple milliseconds. This issue could mess with how enemies spawned which could cause inconsistencies. I fix this issue by integrating parts for spawning waves into the game loop. It is currently unknown if the game loop will break if a second user connects.